requirements.txt
Flask
mysql-connector-python
fuzzywuzzy
python-Levenshtein
gTTS
python-dotenv